refactor(export): extract CSV encoder out of data_exporter (Refs #563 phase: data_exporter)
Move the RFC 4180 CSV primitives (encodeCsv + cell escaping) into a sibling
helper file so data_exporter.dart can stay focused on per-category serialization.
This brings data_exporter.dart from 308 to 280 LOC, satisfying the
<300-LOC objective for the file-size cleanup.

No behavior change. All existing test/core/export/ tests pass.
